+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## Unreleased
 
+* Fix Qgis 3.30 supports
+
 ## 1.8.5 - 2023-03-28
 
 * Use quote format for layers parameter name in output urls